entry endpoint

+++ src/db.rs
@@ -64,7 +64,7 @@ pub async fn connect() -> sqlx::PgPool {
         CREATE TABLE IF NOT EXISTS entry (
             id          SERIAL PRIMARY KEY,
             title       VARCHAR NOT NULL,
-            amount      integer NOT NULL,
+            amount      NUMERIC NOT NULL,
             category    integer,
             created_by  integer NOT NULL,
             note        VARCHAR,
